Drops ListUtil class
