--- app/src/main/java/io/github/alansanchezp/gnomy/ui/ISpinnerItem.java
@@ -1,9 +1,0 @@
-package io.github.alansanchezp.gnomy.ui;
-
-import androidx.annotation.Nullable;
-
-public interface ISpinnerItem {
-    @Nullable String getDrawableResourceName();
-    int getDrawableColor();
-    int getId();
-}
